Add working VGA buffer and testing framework
The VGA buffer plugs into the Rust's println!() implementation.
This allows us to write to the screen in a very rudimentary
way.

The testing framework is pretty basic at this point, but it can
run tests and return results! Test results are passed back to the
host through a serial connection from QEMU, printed to stdout.