interrupt: Provide initial implementation.
The initial implementation of interrupts and exceptions takes a
bit of a shortcut by using the x86_64 crate and the unstable
x86-interrupt function ABI interface.

The x86_64 package provides an implementation of the interrupt
descriptor table. We can provide a handler function reference
to an implementation of the exception handler for that exception.