Interrupts: Support all double page faults with gdt.
Implement a global descriptor table, which is an archaic way to
change stacks when handling some classes of exceptions. We need to
do this because the stack pointer can point to bad memory, causing
a triple page fault. We need to point to a known good stack, which
we can define with the GDT table with a task state segment.