Paging: Implement page fault handler.
The page fault handler will print a debug message to VGA detailing the
offending address. Not very complicated at the moment, however this will
be used in the future to implement writing out pages to spinning rust
when physical RAM is exhausted.

This is a pretty nice abstraction normally, however this is required
when writing an x86_64 operating system.